refactor(messages): specs file, test and story adapted - TWIG-149 (#250)
