+++ php/php_storybook/resources/fix-data.js
@@ -7,8 +7,6 @@ function fixDataInput(data, componentRootName) {
     data = data.dataWithTranslation;
   } else if (data.bannerDataDefault) {
     data = data.bannerDataDefault;
-  } else if (data.dataInfo) {
-    data = data.dataInfo;
   } else if (data.dataGroup1) {
     data = data.dataGroup1;
   } else if (data.englishData) {
