+++ src/ec/packages/ec-component-message/demo/data--error.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-message/demo/data--error';
+import adapter from '../adapter';
+
+export default adapter(specs);
