+++ src/ec/packages/ec-component-message/demo/data--success.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-message/demo/data--success';
+import adapter from '../adapter';
+
+export default adapter(specs);
