+++ src/ec/packages/ec-component-message/demo/data--warning.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-message/demo/data--warning';
+import adapter from '../adapter';
+
+export default adapter(specs);
